camera/requirements.txt
appdirs==1.4.4
attrs==21.4.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorzero==2.0
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.63.0
ffmpeg-python==0.2.0
flake8==3.9.0
future==0.18.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.2
packaging==21.3
pathspec==0.8.1
picamerax==20.9.1
Pillow==8.2.0
pluggy==1.0.0
priority==1.3.0
py==1.11.0
pycodestyle==2.7.0
pydantic==1.8.1
pyflakes==2.3.1
pyparsing==3.0.6
pytest==6.2.5
regex==2021.4.4
requests==2.25.1
scipy==1.6.2
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wsproto==1.0.0
zipp==3.4.1
